reorder user @types
